assign apartment to renter by socialnumber

+++ src/main/java/com/group5/dreamapartment/service/ApartmentService.java
@@ -47,6 +47,10 @@ public class ApartmentService {
   }
 
   public void deleteById(Long id) {
+    Apartment apt = this.findApt(id);
+    if (apt.getRenter() != null) {
+      throw new IllegalStateException("Apartment is not vacant, thus cannot be deleted");
+    }
     apartmentRepository.deleteById(id);
   }
